# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Quản Lý Cân Xe API",
    "version": "1.0.0",
    "description": "API tài liệu cho hệ thống quản lý cân xe.\n\nHệ thống này cung cấp API để quản lý việc cân xe, thông tin khách hàng, \nhàng hóa và các chức năng liên quan đến quản lý phiếu cân.\n",
    "license": {
      "name": "Licensed Under MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Your Name",
      "url": "https://yourwebsite.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Form": {
        "type": "object",
        "properties": {
          "formId": {
            "type": "number",
            "description": "ID form"
          },
          "ten": {
            "type": "string",
            "description": "Tên form"
          },
          "form": {
            "type": "string",
            "description": "Tên file form"
          },
          "menuname": {
            "type": "string",
            "description": "Tên menu"
          },
          "vitri": {
            "type": "string",
            "description": "Vị trí menu"
          }
        },
        "example": {
          "formId": 1,
          "ten": "Thông số hệ thống",
          "form": "frmCaidat",
          "menuname": "mnuThongsohethong",
          "vitri": "danhmuc"
        }
      },
...
      "Hanghoa": {
        "type": "object",
        "required": [
          "ma",
          "ten"
        ],
        "properties": {
          "id": {
            "type": "number",
            "description": "ID tự động tạo của hàng hóa"
          },
          "ma": {
            "type": "string",
            "description": "Mã hàng hóa"
          },
          "ten": {
            "type": "string",
            "description": "Tên hàng hóa"
          },
          "dongia": {
            "type": "number",
            "description": "Đơn giá của hàng hóa"
          }
        },
        "example": {
          "id": 1,
          "ma": "H01",
          "ten": "Cát xây dựng",
          "dongia": 200000
        }
      },
      "HanghoaCreate": {
        "type": "object",
        "required": [
          "ma",
          "ten"
        ],
        "properties": {
          "ma": {
            "type": "string",
            "description": "Mã hàng hóa"
          },
          "ten": {
            "type": "string",
            "description": "Tên hàng hóa"
          },
          "dongia": {
            "type": "number",
            "description": "Đơn giá của hàng hóa"
          }
        },
        "example": {
          "ma": "H01",
          "ten": "Cát xây dựng",
          "dongia": 200000
        }
      },
      "HanghoaUpdate": {
        "type": "object",
        "properties": {
          "ma": {
            "type": "string",
            "description": "Mã hàng hóa"
          },
          "ten": {
            "type": "string",
            "description": "Tên hàng hóa"
          },
          "dongia": {
            "type": "number",
            "description": "Đơn giá của hàng hóa"
          }
        },
        "example": {
          "ten": "Cát xây dựng mịn",
          "dongia": 220000
        }
      },
...
      "NhomQuyen": {
        "type": "object",
        "required": [
          "ma",
          "ten"
        ],
        "properties": {
          "nhomId": {
            "type": "number",
            "description": "ID tự động tạo của nhóm quyền"
          },
          "ma": {
            "type": "string",
            "description": "Mã nhóm quyền"
          },
          "ten": {
            "type": "string",
            "description": "Tên nhóm quyền"
          }
        },
        "example": {
          "nhomId": 1,
          "ma": "admin",
          "ten": "Admin"
        }
      },
...
      "Quyen": {
        "type": "object",
        "required": [
          "nhomId",
          "formId"
        ],
        "properties": {
          "quyenId": {
            "type": "number",
            "description": "ID tự động tạo của quyền"
          },
          "nhomId": {
            "type": "number",
            "description": "ID nhóm quyền"
          },
          "formId": {
            "type": "number",
            "description": "ID form"
          },
          "nhomQuyen": {
            "$ref": "#/components/schemas/NhomQuyen"
          },
          "form": {
            "$ref": "#/components/schemas/Form"
          }
        },
        "example": {
          "quyenId": 1,
          "nhomId": 1,
          "formId": 1,
          "form": {
            "formId": 1,
            "ten": "Thông số hệ thống"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/auth/validate-token": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Kiểm tra tính hợp lệ của token",
        "description": "Kiểm tra xem token hiện tại có hợp lệ hay không",
        "operationId": "validateToken",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Token hợp lệ"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "valid": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Đổi mật khẩu",
        "description": "Đổi mật khẩu của người dùng hiện tại",
        "operationId": "changePassword",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "description": "Mật khẩu hiện tại"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "Mật khẩu mới"
                  }
                },
                "example": {
                  "currentPassword": "password123",
                  "newPassword": "newpassword456"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đổi mật khẩu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Đổi mật khẩu thành công"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/permissions": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Lấy quyền của người dùng",
        "description": "Lấy danh sách quyền của người dùng hiện tại",
        "operationId": "getUserPermissions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách quyền",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Quyền người dùng"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Quyen"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/hanghoa": {
      "get": {
        "tags": [
          "Hàng Hóa"
        ],
        "summary": "Lấy danh sách hàng hóa",
        "description": "Lấy toàn bộ danh sách hàng hóa trong hệ thống",
        "operationId": "getAllHanghoas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách hàng hóa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Danh sách hàng hóa"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Hanghoa"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Hàng Hóa"
        ],
        "summary": "Tạo hàng hóa mới",
        "description": "Tạo một mặt hàng mới trong hệ thống",
        "operationId": "createHanghoa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HanghoaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo hàng hóa thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo hàng hóa thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Hanghoa"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/hanghoa/search": {
      "get": {
        "tags": [
          "Hàng Hóa"
        ],
        "summary": "Tìm kiếm hàng hóa theo tên",
        "description": "Tìm kiếm hàng hóa theo tên hoặc trả về tất cả nếu không có từ khóa",
        "operationId": "searchHanghoaByName",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Từ khóa tìm kiếm tên hàng hóa"
          }
        ],
        "responses": {
          "200": {
            "description": "Kết quả tìm kiếm",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Kết quả tìm kiếm"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Hanghoa"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/hanghoa/{id}": {
      "get": {
        "tags": [
          "Hàng Hóa"
        ],
        "summary": "Lấy thông tin hàng hóa theo ID",
        "description": "Lấy chi tiết một hàng hóa dựa trên ID",
        "operationId": "getHanghoaById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết hàng hóa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Chi tiết hàng hóa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Hanghoa"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "Hàng Hóa"
        ],
        "summary": "Cập nhật hàng hóa",
        "description": "Cập nhật thông tin hàng hóa dựa trên ID",
        "operationId": "updateHanghoa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HanghoaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật hàng hóa thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Cập nhật hàng hóa thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Hanghoa"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Hàng Hóa"
        ],
        "summary": "Xóa hàng hóa",
        "description": "Xóa một hàng hóa dựa trên ID",
        "operationId": "deleteHanghoa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa hàng hóa thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Xóa hàng hóa thành công"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/hanghoa/code/{ma}": {
      "get": {
        "tags": [
          "Hàng Hóa"
        ],
        "summary": "Lấy thông tin hàng hóa theo mã",
        "description": "Lấy chi tiết một hàng hóa dựa trên mã",
        "operationId": "getHanghoaByMa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ma",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Mã hàng hóa"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết hàng hóa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Chi tiết hàng hóa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Hanghoa"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}